+++ common.blocks/i-bem/__i18n/i-bem__i18n.test.bemhtml/files/lang.json
